+++ CHANGES.rst
@@ -68,6 +68,7 @@ v3.0.0 (xxxx-xx-xx)
   - ``surround``
 
 - Make ``capitalize`` accept ``lower_rest`` argument to control whether to convert the rest of the string to lower case or not. Defaults to ``True``.
+- Make ``chain`` support late passing of initial ``value`` argument.
 - Make ``splice`` work with strings.
 - Make ``to_string`` convert ``None`` to empty string. (**breaking change**)
 - Make ``re_replace`` accept non-stirng arguments.